cplay: support auto-configuration of fragment size and count
Default the fragment size and count to zero to take advantage
of the auto-configuration support that was added to tinycompress.
If user doesn't override these by specifying -f and -b values
on the command line, they will be passed to tinycompress as zero
and tinycompress will query the driver for size and count.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>