wave: Introduce parse_wave header
This parses a canonical WAVE header and fills in
informationlike channels, rate and format.

In tinycompress, we use canonical WAVE format described here:

http://soundfile.sapp.org/doc/WaveFormat/

A WAVE file is often just a RIFF file with a single "WAVE" chunk which
consists of two sub-chunks -- a "fmt " chunk specifying the data format
and a "data" chunk containing the actual sample data.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>